Fix `bail!` function references and migrate tests from unit into integrational
