+++ gstd/src/lib.rs
@@ -23,7 +23,7 @@
 
 extern crate galloc;
 
-pub mod bail;
+mod bail;
 #[cfg(feature = "meta")]
 pub mod meta;
 pub mod msg;
